Add annotations to add DNS for Jenkins
These annotations should ensure that the DNS record gets added when the
service is created and points to the ELB created.

The thing that picks up the annotations and ensures that the DNS record
is in place is the "external-dns" Deployment, running in the "default"
namespace.

This looks to be based off of the:

  registry.opensource.zalan.do/teapot/external-dns:v0.4.5

Docker image.

Issue(s): None